+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project made by Monade Team are documented in this file. For info refer to team@monade.io
 
+## [0.2.0] - 2020-07-02
+
+### Added
+- Support for JSON:API-style pagination parameters
+
+### Fixed
+- Broken dependency with kaminari/activerecord
+
 ## [0.1.0] - 2020-03-20
 First functional version